requirements.txt
fastapi
uvicorn
pytesseract
pillow
pdf2image
pyzbar
pyaadhaar==2.0.2
python-multipart